requirements .txt
OpenAI
streamlit
python-dotenv
os
